Tools/requirements.txt
pyserial
prompt_toolkit
